lab7/two_stage_protection/requirements.txt
cryptography==41.0.7
pillow==10.1.0
numpy==1.24.3
